AI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6 
aiortc==1.5.0
httpx==0.27.0
python-socketio==5.11.2
aiohttp>=3.8.0
twilio>=8.10.0

numpy==1.26.1
soundfile==0.12.1
ctranslate2==3.23.0
faster-whisper==0.10.1
huggingface_hub[hf_xet]>=0.20.0
scipy>=1.10.0


opencv-python==4.8.1.78
ultralytics
torch==2.0.1
torchvision==0.15.2
lap>=0.5.12